+++ twilight-http/src/request/application/interaction/update_response.rs
@@ -46,8 +46,8 @@ struct UpdateResponseFields<'a> {
 ///
 /// You can pass [`None`] to any of the methods to remove the associated field.
 /// Pass [`None`] to [`content`] to remove the content. You must ensure that the
-/// message still contains at least one of [`attachments`], [`content`], or
-/// [`embeds`].
+/// message still contains at least one of [`attachments`], [`components`],
+/// [`content`] or [`embeds`].
 ///
 /// This endpoint is not bound to the application's global rate limit.
 ///
@@ -79,6 +79,7 @@ struct UpdateResponseFields<'a> {
 /// ```
 ///
 /// [`attachments`]: Self::attachments
+/// [`components`]: Self::components
 /// [`content`]: Self::content
 /// [`embeds`]: Self::embeds
 #[must_use = "requests must be configured and executed"]