+++ twilight-http/src/request/channel/invite/create_invite.rs
@@ -118,6 +118,7 @@ impl<'a> CreateInvite<'a> {
     ///
     /// [`InviteMaxAge`]: twilight_validate::request::ValidationErrorType::InviteMaxAge
     pub const fn max_age(mut self, max_age: u32) -> Result<Self, ValidationError> {
+        #[allow(clippy::question_mark)]
         if let Err(source) = validate_invite_max_age(max_age) {
             return Err(source);
         }
@@ -161,6 +162,7 @@ impl<'a> CreateInvite<'a> {
     ///
     /// [`InviteMaxUses`]: twilight_validate::request::ValidationErrorType::InviteMaxUses
     pub const fn max_uses(mut self, max_uses: u16) -> Result<Self, ValidationError> {
+        #[allow(clippy::question_mark)]
         if let Err(source) = validate_invite_max_uses(max_uses) {
             return Err(source);
         }