+++ twilight-http/src/request/channel/update_channel.rs
@@ -94,6 +94,7 @@ impl<'a> UpdateChannel<'a> {
     ///
     /// [`BitrateInvalid`]: twilight_validate::channel::ChannelValidationErrorType::BitrateInvalid
     pub const fn bitrate(mut self, bitrate: u32) -> Result<Self, ChannelValidationError> {
+        #[allow(clippy::question_mark)]
         if let Err(source) = validate_bitrate(bitrate) {
             return Err(source);
         }
@@ -174,6 +175,7 @@ impl<'a> UpdateChannel<'a> {
         mut self,
         rate_limit_per_user: u16,
     ) -> Result<Self, ChannelValidationError> {
+        #[allow(clippy::question_mark)]
         if let Err(source) = twilight_validate::channel::rate_limit_per_user(rate_limit_per_user) {
             return Err(source);
         }